feat(eslint): enable no-unnecessary-type-assertion and strip redundant casts (#3668)
## Summary

- Adds a scoped typed-lint config block (`projectService: true`,
`tsconfigRootDir`) to `eslint.config.mjs` that wires the TypeScript
program into the ESLint parser for `**/*.ts` files and enables
`@typescript-eslint/no-unnecessary-type-assertion: "error"`.
- Ran `eslint --fix` to auto-strip provably-redundant type assertions
across the codebase.
- Restored one removal in `migration.ts` (~line 2975) where `tseslint`'s
narrowing analysis diverged from `tsc` — the `as "up" | "down"` cast is
genuinely required to satisfy the function return type and is suppressed
inline.

## LOC note (exceeds 500 LOC ceiling)

This is a **single mechanical autofix** across 397 files (~3.5k LOC).
The diff is large by line count but structurally trivial: every change
is the deletion of a redundant cast or non-null assertion that
TypeScript's own type-checker proved unnecessary. No logic was changed;
the fixer only removes dead cast syntax.

## CI / lint performance

`projectService: true` adds a one-time TypeScript program initialization
to the `eslint` process (~seconds on a cold run). It is **not** the same
as running `tsc` again — the TS language service is shared across all
files in the same lint invocation. The block is kept narrowly scoped
(one rule, one block) rather than switching the whole config to
`recommendedTypeChecked`, so no other type-checked rules are activated.
If the lint job's wall-clock time becomes a concern, this block can be
extracted to a separate `lint:typed` script; for now it's in the default
path so CI enforces the rule on every push.

## Test plan

- [x] `pnpm build` passes (no `tsc --build` errors after restoring the
`migration.ts` cast)
- [ ] CI lint job passes
- [ ] CI tests pass (no behavioral changes expected — only dead cast
syntax removed)

+++ packages/activemodel/src/type/binary.ts
@@ -30,9 +30,9 @@ export class BinaryType extends ValueType<Uint8Array> {
     const cur = this.serialize(newValue);
     if (old === null && cur === null) return false;
     if (old === null || cur === null) return true;
-    if (old.length !== (cur as Uint8Array).length) return true;
+    if (old.length !== cur.length) return true;
     for (let i = 0; i < old.length; i++) {
-      if (old[i] !== (cur as Uint8Array)[i]) return true;
+      if (old[i] !== cur[i]) return true;
     }
     return false;
   }
